Posts

+++ SocialNetwork/domain/src/main/java/ru/skillbox/dto/PostDto.java
@@ -4,19 +4,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 import ru.skillbox.dto.enums.Type;
-import ru.skillbox.model.Like;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import java.util.List;
 
 @Getter
 @Setter
 public class PostDto {
 
-    private Integer id;
+    private Long id;
     private Long time;
     private AccountByIdDto author;
     private String title;
